[refactor]: Altera o nome do arquivo de configuração do banco de dados de testes
O arquivo do banco de dados de testes possuia um nome sem anexo com o de configuração da aplicação de produção
